# api/src/openapi.yaml
openapi: 3.1.0
info:
  title: Deffavia API
  version: "1.1.0"
  description: API para reservas de asientos (orden + ítems), exportación/importación XML, reportes y correo.
servers:
  - url: http://localhost:3001
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        ok: { type: boolean, example: false }
        error: { type: string }
...
    CreateSelection:
      type: object
      required: [seatCode, passengerName, cui, hasLuggage]
      properties:
        seatCode: { type: string, example: E5 }
        passengerName: { type: string, example: Juan Pérez }
        cui: { type: string, example: "1234567890123" }
        hasLuggage: { type: boolean, example: true }
    CreateOrderPayloadManual:
      type: object
      required: [userEmail, mode, selections]
      properties:
        userEmail: { type: string, format: email }
        mode: { type: string, enum: [manual], example: manual }
        selections:
          type: array
          items: { $ref: '#/components/schemas/CreateSelection' }
    CreateOrderPayloadRandom:
      type: object
      required: [userEmail, mode, random]
      properties:
        userEmail: { type: string, format: email }
        mode: { type: string, enum: [random], example: random }
        random:
          type: object
          required: [seatClass, count, passengers]
          properties:
            seatClass: { type: string, enum: [business, economy] }
            count: { type: integer, minimum: 1 }
            passengers:
              type: array
              items:
                type: object
                required: [passengerName, cui, hasLuggage]
                properties:
                  passengerName: { type: string }
                  cui: { type: string }
                  hasLuggage: { type: boolean }
    CreateOrderPayloadGroup:
      type: object
      required: [userEmail, mode, group]
      properties:
        userEmail: { type: string, format: email }
        mode: { type: string, enum: [group], example: group }
        group:
          type: object
          required: [seatClass, count, passengers]
          properties:
            seatClass: { type: string, enum: [business, economy] }
            count: { type: integer, minimum: 2 }
            requireContiguous: { type: boolean, default: true }
            groupName: { type: string, nullable: true }
            passengers:
              type: array
              minItems: 2
              items:
                type: object
                required: [passengerName, cui, hasLuggage]
                properties:
                  passengerName: { type: string }
                  cui: { type: string }
                  hasLuggage: { type: boolean }
                  isLeader: { type: boolean, description: "Marca líder del grupo (opcional, por defecto el primero)" }
    CreatedOrderResponse:
      type: object
      properties:
        ok: { type: boolean }
        orderId: { type: integer }
        isVip: { type: boolean }
        subtotal: { type: number }
        discountTotal: { type: number }
        total: { type: number }
        items:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              seatCode: { type: string }
              seatClass: { type: string, enum: [business, economy] }
              passengerName: { type: string }
              cui: { type: string }
              total: { type: number }
    ModifiedItemSeatResponse:
      type: object
      properties:
        ok: { type: boolean }
        increment: { type: number, description: "+10% aplicado" }
        from: { type: string, description: "Código de asiento anterior" }
        to: { type: string, description: "Nuevo código de asiento" }
        oldItemTotal: { type: number }
        newItemTotal: { type: number }
        orderTotal: { type: number }
...
paths:
...
  /api/reservations:
    post:
      tags: [Reservations]
      summary: Crear reserva (manual, aleatoria o grupal)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/CreateOrderPayloadManual'
                - $ref: '#/components/schemas/CreateOrderPayloadRandom'
                - $ref: '#/components/schemas/CreateOrderPayloadGroup'
      responses:
        '200':
          description: Creada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedOrderResponse' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/reservations/my:
    get:
      tags: [Reservations]
      summary: Mis reservas (usuario autenticado)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /api/reservations/{orderId}/items/{itemId}/seat:
    patch:
      tags: [Reservations]
      summary: Modificar asiento (+10%)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: integer }
        - in: path
          name: itemId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newSeatCode, cui]
              properties:
                newSeatCode: { type: string }
                cui: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModifiedItemSeatResponse' }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /api/reservations/{orderId}/items/{itemId}/cancel:
    post:
      tags: [Reservations]
      summary: Cancelar asiento
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: integer }
        - in: path
          name: itemId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cui]
              properties:
                cui: { type: string }
      responses:
        '200': { description: OK }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/reservations/{orderId}/cancel:
    post:
      tags: [Reservations]
      summary: Cancelar una orden completa (marca ítems activos como cancelados)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK }
        '400':
          description: Error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...